DROID-3905 fix
