+++ app/src/main/java/com/anytypeio/anytype/ui/vault/VaultSpaceActions.kt
@@ -1,9 +1,7 @@
 package com.anytypeio.anytype.ui.vault
 
 import androidx.compose.foundation.Image
-import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxSize
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
@@ -12,16 +10,11 @@ import androidx.compose.material.Text
 import androidx.compose.material3.DropdownMenu
 import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.runtime.Composable
-import androidx.compose.runtime.getValue
-import androidx.compose.runtime.mutableStateOf
-import androidx.compose.runtime.remember
-import androidx.compose.runtime.setValue
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.colorResource
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.DpOffset
 import androidx.compose.ui.unit.dp
 import com.anytypeio.anytype.R
@@ -57,12 +50,15 @@ fun SpaceActionsDropdownMenu(
     ) {
         // Show pin/unpin option or limit message
         val canPin = currentPinnedCount < VaultUiState.MAX_PINNED_SPACES
-        
+
         if (!isPinned && !canPin) {
             // Show info message when limit reached and space is not pinned
             Row(modifier = Modifier.padding(16.dp)) {
                 Text(
-                    text = stringResource(R.string.vault_pinned_limit_message, VaultUiState.MAX_PINNED_SPACES),
+                    text = stringResource(
+                        R.string.vault_pinned_limit_message,
+                        VaultUiState.MAX_PINNED_SPACES
+                    ),
                     style = PreviewTitle2Regular,
                     color = colorResource(id = R.color.text_secondary)
                 )
@@ -155,42 +151,6 @@ fun SpaceActionsDropdownMenu(
                     )
                 }
             }
-        )
-    }
-}
-
-@Preview(showBackground = true, name = "SpaceActionsDropdownMenu - Muted Owner")
-@Composable
-fun PreviewSpaceActionsDropdownMenu_MutedOwner() {
-    var expanded by remember { mutableStateOf(true) }
-    Box(Modifier.fillMaxSize()) {
-        SpaceActionsDropdownMenu(
-            expanded = expanded,
-            onDismiss = { expanded = false },
-            isMuted = true,
-            isPinned = false,
-            currentPinnedCount = 3,
-            onMuteToggle = {},
-            onPinToggle = {},
-            onSpaceSettings = {}
-        )
-    }
-}
-
-@Preview(showBackground = true, name = "SpaceActionsDropdownMenu - Unmuted Not Owner")
-@Composable
-fun PreviewSpaceActionsDropdownMenu_UnmutedNotOwner() {
-    var expanded by remember { mutableStateOf(true) }
-    Box(Modifier.fillMaxSize()) {
-        SpaceActionsDropdownMenu(
-            expanded = expanded,
-            onDismiss = { expanded = false },
-            isMuted = false,
-            isPinned = false,
-            currentPinnedCount = 6,
-            onMuteToggle = {},
-            onPinToggle = {},
-            onSpaceSettings = {}
         )
     }
 }
